feat(comboBox): move agreement search & company search ti use comboBox component which accessible
