+++ packages/code-du-travail-frontend/src/outils/ConventionCollective/steps/EnterpriseSearch.tsx
@@ -1,10 +1,11 @@
 import { SOURCES } from "@socialgouv/cdtn-sources";
 import {
   Button,
+  ComboBoxList,
+  ComboBoxOption,
+  ComboBoxPopover,
   ScreenReaderOnly,
   Section as SectionUi,
-  Title,
-  Wrapper,
 } from "@socialgouv/cdtn-ui";
 import Link from "next/link";
 import React, { MutableRefObject, useRef } from "react";
@@ -13,11 +14,9 @@ import styled from "styled-components";
 
 import { Enterprise } from "../../../conventions/Search/api/enterprises.service";
 import { InlineError } from "../../common/ErrorField";
-import { HelpModal } from "../common/Modal";
 import { useNavContext } from "../common/NavContext";
-import { ListItem, ResultList } from "../common/ResultList";
-import { useTrackingContext } from "../common/TrackingContext";
 import { EnterpriseButton } from "../enterprise/EnterpriseButton";
+import NoResultSection from "../enterprise/NoResultSection";
 import { SearchEnterprise, SearchParams } from "../enterprise/SearchEnterprise";
 
 type EnterpriseSearchStepProps = {
@@ -28,7 +27,6 @@ const EnterpriseSearchStep = ({
   onBackClick,
 }: EnterpriseSearchStepProps): JSX.Element => {
   const { setSearchParams, setEnterprise } = useNavContext();
-  const { trackEvent, uuid, title } = useTrackingContext();
   const refInput = useRef<HTMLDivElement>();
 
   function handleEnterpriseSelection(
@@ -37,11 +35,6 @@ const EnterpriseSearchStep = ({
   ) {
     setEnterprise(enterprise);
     setSearchParams(params);
-  }
-
-  function openModalHandler(openModal: () => void) {
-    trackEvent("cc_search_help", "click_cc_search_help_p2", title, uuid);
-    openModal();
   }
 
   return (
@@ -72,16 +65,20 @@ const EnterpriseSearchStep = ({
             return <Section role="status">{state.error}</Section>;
           }
 
-          return state.data ? (
-            state.data.length > 0 ? (
-              <Section>
-                <ScreenReaderOnly role="status">
-                  {state.data.length} résultats
-                </ScreenReaderOnly>
-                <ResultList query={`${params.query}-${params.address}`}>
+          if (!state.data) {
+            return <></>;
+          }
+          return state.data.length > 0 ? (
+            <Section>
+              <ScreenReaderOnly role="status">
+                {state.data.length} résultats
+              </ScreenReaderOnly>
+
+              <ComboBoxPopover>
+                <ComboBoxList>
                   {state.data.map((item, index) => {
                     return (
-                      <ListItem key={item.siren}>
+                      <ComboBoxOption value={item.simpleLabel} key={item.siren}>
                         <EnterpriseButton
                           showAddress={params.address.length > 0 || isSiret}
                           isFirst={index === 0}
@@ -90,66 +87,14 @@ const EnterpriseSearchStep = ({
                             handleEnterpriseSelection(item, params)
                           }
                         />
-                      </ListItem>
+                      </ComboBoxOption>
                     );
                   })}
-                </ResultList>
-              </Section>
-            ) : (
-              <Section>
-                <Wrapper variant="light">
-                  <p>
-                    <strong>Aucune entreprise n’a été trouvée</strong>.
-                  </p>
-                  Suggestions&nbsp;:
-                  <ul>
-                    <li>Vérifiez l’orthographe des termes de recherche</li>
-                    <li>
-                      Utilisez la rubrique ci-dessous “Vous ne trouvez pas votre
-                      entreprise&nbsp;? Consultez notre aide”
-                    </li>
-                  </ul>
-                </Wrapper>
-                <Wrapper>
-                  Vous ne trouvez pas votre entreprise&nbsp;?
-                  <br />
-                  <HelpModal
-                    title="Vous ne trouvez pas votre convention collective"
-                    renderButton={(openModal) => (
-                      <Button
-                        variant="link"
-                        onClick={() => openModalHandler(openModal)}
-                      >
-                        Consulter notre aide
-                      </Button>
-                    )}
-                  >
-                    <Title stripe="none" as="h3">
-                      Vous ne trouvez pas votre convention collective&nbsp;?
-                    </Title>
-                    <p>Il peut y avoir plusieurs explications à cela&nbsp;:</p>
-                    <ul>
-                      <li>
-                        Votre entreprise a été enregistrée sous un autre nom ou
-                        un autre code : si vous le pouvez, utilisez son numéro
-                        Siret. Ce dernier doit être présent sur votre bulletin
-                        de paie.
-                      </li>
-                      <li>
-                        Votre entreprise a un statut particulier :
-                        administration ou établissements publics, associations,
-                        secteur agricole, La Poste, La Croix Rouge etc. ;
-                      </li>
-                      <li>
-                        Votre entreprise n’a pas de convention collective.
-                      </li>
-                    </ul>
-                  </HelpModal>
-                </Wrapper>
-              </Section>
-            )
+                </ComboBoxList>
+              </ComboBoxPopover>
+            </Section>
           ) : (
-            <></>
+            <NoResultSection />
           );
         }}
       />
