+++ packages/react-ui/src/field/ComboBox.js
@@ -9,7 +9,7 @@ import React from "react";
 import styled from "styled-components";
 
 import { Search } from "../icons/index.js";
-import { box, fonts, input, spacings } from "../theme";
+import { box, colors, fonts, input, spacings } from "../theme";
 
 export const ComboBox = ({ ...props }) => <Combobox {...props} />;
 
@@ -26,14 +26,14 @@ const StyledComboBoxInput = styled(({ ...props }) => (
   <ComboboxInput {...props} />
 ))`
   height: ${input.height};
-  padding: 0 5rem 0 ${spacings.medium};
+  width: 100%;
+  padding: 0 4rem 0 ${spacings.medium};
   color: ${({ theme }) => theme.paragraph};
   font-size: ${fonts.sizes.default};
   font-family: "Open Sans", sans-serif;
   background: ${({ theme }) => theme.white};
-  border: 1px solid transparent;
-  border-color: ${({ invalid, theme }) =>
-    invalid ? theme.error : "transparent"};
+  border: 1px solid
+    ${({ invalid, theme }) => (invalid ? theme.error : "transparent")};
   border-radius: ${box.borderRadius};
   box-shadow: ${({ theme }) => box.shadow.default(theme.secondary)};
 `;
@@ -56,8 +56,11 @@ export const ComboBoxPopover = styled(({ ...props }) => (
 ))`
   border-radius: ${box.borderRadius};
   color: ${({ theme }) => theme.paragraph};
-  border: solid 1px hsla(0, 0%, 0%, 0.25);
-  background: hsla(0, 100%, 100%, 0.99);
+  background: ${colors.bgPrimary};
+  border: ${({ theme }) => box.border(theme.border)};
+  max-height: 400px;
+  overflow-y: auto;
+  margin-top: ${spacings.small};
 `;
 
 export const ComboBoxOption = styled(({ ...props }) => (
@@ -67,16 +70,13 @@ export const ComboBoxOption = styled(({ ...props }) => (
   font-family: "Open Sans", sans-serif;
   cursor: pointer;
 
-  [aria-selected="true"] {
-    background: hsl(211, 10%, 95%);
+  &:hover,
+  &[aria-selected="true"] {
+    background: ${colors.bgTertiary};
   }
 
-  :hover {
-    background: hsl(211, 10%, 92%);
-  }
-
-  [aria-selected="true"]:hover {
-    background: hsl(211, 10%, 90%);
+  & + & {
+    border-top: ${({ theme }) => box.border(theme.border)};
   }
 `;
